+++ src/auth/auth.module.ts
@@ -2,11 +2,9 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { ConfigModule } from '@nestjs/config';
-import { KakaoStrategy } from 'src/util/strategy/kakao.strategy';
 import jwtConfig from 'src/configs/auth/jwt.config';
 import { JwtModule } from '@nestjs/jwt';
 import refreshJwtConfig from 'src/configs/auth/refresh.jwt.config';
-import { UsersService } from 'src/user/user.service';
 import { User } from '../entity/user.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { JwtStrategy } from 'src/util/strategy/jwt.strategy';
@@ -14,9 +12,6 @@ import { RefreshJwtStrategy } from 'src/util/strategy/refresh.strategy';
 import { APP_GUARD } from '@nestjs/core';
 import { JwtAuthGuard } from 'src/util/guards/jwt-auth/jwt-auth.guard';
 import { UsersModule } from 'src/user/user.module';
-import { HttpModule } from '@nestjs/axios';
-import oauthConfig from 'src/configs/auth/oauth.config';
-import kakaoOauthConfig from 'src/configs/auth/kakao.oauth.config';
 
 @Module({
   imports: [
@@ -24,17 +19,11 @@ import kakaoOauthConfig from 'src/configs/auth/kakao.oauth.config';
     JwtModule.registerAsync(jwtConfig.asProvider()),
     ConfigModule.forFeature(jwtConfig),
     ConfigModule.forFeature(refreshJwtConfig),
-    ConfigModule.forFeature(oauthConfig),
-    ConfigModule.forFeature(kakaoOauthConfig),
-    AuthModule,
     UsersModule,
-    HttpModule,
   ],
   controllers: [AuthController],
   providers: [
     AuthService,
-    UsersService,
-    KakaoStrategy,
     JwtStrategy,
     RefreshJwtStrategy,
     {
@@ -42,6 +31,6 @@ import kakaoOauthConfig from 'src/configs/auth/kakao.oauth.config';
       useClass: JwtAuthGuard,
     },
   ],
-  exports: [AuthService, UsersService],
+  exports: [AuthService, JwtModule],
 })
 export class AuthModule {}
